requirements.txt
bcrypt==4.1.3
fastapi==0.111.0
passlib==1.7.4
PyJWT==2.8.0
uvicorn==0.29.0
pydantic-settings==2.2.1
redis==5.0.4
kafka-python==2.0.2
aiokafka==0.10.0
loguru==0.7.2
charset-normalizer==3.3.2
requests==2.32.3
urllib3==2.2.2
kafka-python-ng==2.2.2
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
frozenlist==1.4.1
multidict==6.0.5
yarl==1.9.4
fastapi-cache2==0.2.1
pendulum==3.0.0
python-dateutil==2.9.0.post0
six==1.16.0
time-machine==2.14.2
tzdata==2024.1
lxml==6.0.2
moviepy==1.0.3
Pillow>=10.0.0
google-auth>=2.0.0
python-multipart>=0.0.6
